Moved FileFtpReference to models
